requirements.txt
tgcrypto
pytz
psutil
motor
dnspython
GitPython
pykeyboard
bs4
Pillow
pymediainfo
opencv-python-headless
requests
unidecode
attrify
pySmartDL
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
telegraph
pyromod==1.1
youtube-search-python
pymongo
yt-dlp
wget
python-dateutil
emoji
python-ffmpeg
search_engine_parser
g4f==0.1.9.3
openai
uvloop
google-generativeai
pytimeparse
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
pytgcalls==3.0.0.dev24
pytz==2024.1
git+https://github.com/naya1503/mypyrogram@dev